Fix: 401 에러 문제 해결
