+++ app/_libs/authService.ts
@@ -1,4 +1,3 @@
-import type { AxiosError, AxiosRequestConfig } from 'axios';
 import { isAxiosError } from 'axios';
 
 import type { LoginFormValues, Response } from '@/_types/authentication';
@@ -12,83 +11,6 @@ interface ErrorResponse {
 }
 
 type PostLogin = (params: LoginFormValues) => Promise<Response>;
-
-let isRefreshing = false;
-let failedQueue: Array<{
-  reject: (reason?: unknown) => void;
-  resolve: (value: string | PromiseLike<string | undefined> | undefined) => void;
-}> = [];
-
-interface CustomAxiosRequestConfig extends AxiosRequestConfig {
-  retry?: boolean;
-}
-
-const processQueue = (error: unknown, token: string | null = null) => {
-  failedQueue.forEach((prom) => {
-    if (error) {
-      prom.reject(error);
-    } else {
-      prom.resolve(token || undefined);
-    }
-  });
-
-  failedQueue = [];
-};
-
-export const tokenRefresh = async (): Promise<{ accessToken: string; refreshToken: string }> => {
-  const response = await axiosInstance.post<Response>('/auth/tokens');
-  const { accessToken, refreshToken } = response.data;
-  setCookie('accessToken', accessToken);
-  setCookie('refreshToken', refreshToken);
-  axiosInstance.defaults.headers.common.Authorization = `Bearer ${accessToken}`;
-  return { accessToken, refreshToken };
-};
-
-axiosInstance.interceptors.response.use(
-  (response) => response,
-  async (error: AxiosError) => {
-    const originalRequest = error.config as CustomAxiosRequestConfig | undefined;
-
-    if (error.response?.status === 401 && originalRequest && !originalRequest.retry) {
-      if (isRefreshing) {
-        return new Promise<string | undefined>((resolve, reject) => {
-          failedQueue.push({ resolve, reject });
-        })
-          .then((token) => {
-            if (originalRequest) {
-              originalRequest.headers = originalRequest.headers || {};
-              originalRequest.headers.Authorization = `Bearer ${token}`;
-              return axiosInstance(originalRequest);
-            }
-            return Promise.reject(new Error('Original request is undefined'));
-          })
-          .catch((err) => Promise.reject(err));
-      }
-
-      originalRequest.retry = true;
-      isRefreshing = true;
-
-      return new Promise<string | undefined>((resolve, reject) => {
-        tokenRefresh()
-          .then(({ accessToken }) => {
-            originalRequest.headers = originalRequest.headers || {};
-            originalRequest.headers.Authorization = `Bearer ${accessToken}`;
-            processQueue(null, accessToken);
-            resolve(axiosInstance(originalRequest));
-          })
-          .catch((err) => {
-            processQueue(err, null);
-            reject(err);
-          })
-          .finally(() => {
-            isRefreshing = false;
-          });
-      });
-    }
-
-    return Promise.reject(error);
-  },
-);
 
 export const postLogin: PostLogin = async ({ email, password }) => {
   try {
